requirements.txt
google-adk[eval]>=1.5.0
google-cloud-aiplatform[adk,agent-engines]>=1.93.0
google-cloud-storage
litellm>=1.50.0
